Changed how you retrieve specific properties off of a generic MapEntity to be a generic function that auto-casts back.

+++ EditorCore/Common/Maps/ShipSpawn.cs
@@ -5,7 +5,7 @@ namespace WEditor.Common.Maps
     public class ShipSpawn : WObject
     {
         public Vector3 Position;
-        public short YRotation;
+        public float YRotation;
         public byte ShipId;
         public byte Unknown1;
     }
